Add minimal LSP support
This adds minimal LSP support for beancount files if the
beancount-language-server is already installed in the $PATH.

I looked at following the existing pattern for installing the
language server from the GitHub repo. However, the Mac and Linux
archives are .tar.xz files which is not supported in the
DownloadedFileType enum.